+++ checkpoint/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Reduce Memory Pressure from long held CPU buffers in Pathways MTC
+- Support explicit devices for Pathways MTC init
 
 ## [0.12.1] - 2026-06-24